+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Reduce the size of the struct returned by `UnicodeCategorySet.iter()` from 8 to 4 bytes and improve performance for
   cases when a few Unicode categories are involved.
 - Reduce allocations inside `query` for cases when a few Unicode categories are involved.
+- Rename `UnicodeCategorySet::add_category` to `UnicodeCategorySet::add` and `UnicodeCategorySet::has_category` to `UnicodeCategorySet::contains`.
 
 ## [0.1.2] - 2023-04-23